\name{EntropyEstimation-package}
\alias{EntropyEstimation-package}
\alias{EntropyEstimation}
\docType{package}
\title{Estimation of Entropy and Related Quantities}
\description{
Contains methods for the estimation of Shannon's entropy, variants of Renyi's entropy, mutual Information, Kullback-Leibler divergence, and generalized Simpson's indices. These estimators have a bias that decays exponentially fast. For more information see Z. Zhang and J. Zhou (2010), Zhang (2012), Zhang (2013), Zhang and Grabchak (2013), Zhang and Grabchak (2014a), Zhang and Grabchak (2014b), and Zhang and Zheng (2014). 
}
\details{
\tabular{ll}{
Package: \tab EntropyEstimation\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2014-01-04\cr
License: \tab GPL3\cr}
}
\author{
Lijuan Cao <lcao@jcsu.edu> and Michael Grabchak <mgrabcha@uncc.edu> %Maintainer: Michael Grabchak <mgrabcha@uncc.edu>
}
\references{
Z. Zhang (2012). Entropy estimation in Turing's' perspective. Neural Computation 24(5), 1368--1389.

Z. Zhang (2013). Asymptotic normality of an entropy estimator with asymptotically decaying bias. IEEE Transactions on Information Theory 59(1), 504--508.

Z. Zhang and M. Grabchak (2013). Bias Adjustment for a Nonparametric Entropy Estimator. Entropy, 15(6), 1999-2011.

Z. Zhang and M. Grabchak (2014a). Entropic representation and estimation of diversity indices. http://arxiv.org/abs/1403.3031.

Z. Zhang and M. Grabchak (2014b). Nonparametric Estimation of Kullback-Leibler Divergence. Neural Computation, 26(11): 2570-2593.

Z. Zhang and L. Zheng (2014). A Mutual Information Estimator with Exponentially Decaying Bias.

Z. Zhang and J. Zhou (2010). Re-parameterization of multinomial distributions and diversity indices. Journal of Statistical Planning and Inference 140(7), 1731-1738.
}

